requirements.txt
accelerate==1.12.0
aiofiles==25.1.0
albucore==0.0.24
albumentations==2.0.8
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
attrs==25.4.0
av==16.1.0
beautifulsoup4==4.14.3
blinker==1.9.0
boto3==1.42.49
botocore==1.42.49
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.10.1
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
dill==0.4.1
distro==1.9.0
doclayout_yolo==0.0.4
fast-langdetect==0.2.5
fasttext-predict==0.9.2.4
filelock==3.23.0
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-retries==0.4.5
huggingface_hub==1.4.1
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
json_repair==0.57.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
langchain-core==1.2.13
langgraph==1.0.8
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.6
langsmith==0.7.3
lazy_loader==0.4
loguru==0.7.3
lxml==6.0.2
magika==1.0.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mineru==2.7.6
mineru_vl_utils==0.1.22
modelscope==1.34.0
mpmath==1.3.0
narwhals==2.16.0
networkx==3.6.1
numpy==2.4.2
omegaconf==2.3.0
onnxruntime==1.20.1
openai==2.21.0
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pandas==2.3.3
pdfminer.six==20260107
pdftext==0.6.3
pillow==12.1.1
polars==1.38.1
polars-runtime-32==1.38.1
protobuf==6.33.5
psutil==7.2.2
py-cpuinfo==9.0.0
pyarrow==23.0.0
pyclipper==1.4.0
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.2
pypdf==6.7.0
pypdfium2==4.30.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
qwen-vl-utils==0.0.14
referencing==0.37.0
regex==2026.1.15
reportlab==4.4.10
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.2
robust-downloader==0.0.2
rpds-py==0.30.0
s3transfer==0.16.0
safetensors==0.7.0
scikit-image==0.26.0
scipy==1.17.0
seaborn==0.13.2
setuptools==82.0.0
shapely==2.1.2
shellingham==1.5.4
simsimd==6.5.12
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.3
streamlit==1.54.0
stringzilla==4.6.0
sympy==1.14.0
tenacity==9.1.4
thop==0.1.1.post2209072238
tifffile==2026.1.28
tokenizers==0.22.2
toml==0.10.2
torch==2.10.0+cpu
torchaudio==2.10.0+cpu
torchvision==0.25.0+cpu
tornado==6.5.4
tqdm==4.67.3
typer==0.23.1
typer-slim==0.23.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.4.14
ultralytics-thop==2.0.18
urllib3==2.6.3
uuid_utils==0.14.0
uv==0.10.2
watchdog==6.0.0
win32_setctime==1.2.0
xxhash==3.6.0
zstandard==0.25.0
